[refactor] 이미지 업로드 기능을 useChangeProfileImage 훅으로 분리
